+++ src/Deriver/Decisions/Finders/IDecisionFinder.cs
@@ -7,28 +7,3 @@ public interface IDecisionFinder
     bool CanFindDecision(ImportPreNotification notification, CheckCode? checkCode);
     DecisionFinderResult FindDecision(ImportPreNotification notification, CheckCode? checkCode);
 }
-
-public class CheckCode
-{
-    public const string IuuCheckCode = "H224";
-
-    public required string Value { get; set; }
-
-    public ImportNotificationType? GetImportNotificationType()
-    {
-        //This is the mapping from https://eaflood.atlassian.net/wiki/spaces/ALVS/pages/5400920093/DocumentCode+CheckCode+Mapping
-        return Value switch
-        {
-            "H221" => ImportNotificationType.Cveda,
-            "H223" => ImportNotificationType.Ced,
-            "H222" or "H224" => ImportNotificationType.Cvedp,
-            "H219" or "H218" or "H220" => ImportNotificationType.Chedpp,
-            _ => null,
-        };
-    }
-
-    public bool IsIuu()
-    {
-        return Value == IuuCheckCode;
-    }
-}
