+++ src/Deriver/Extensions/ServiceCollectionExtensions.cs
@@ -1,5 +1,4 @@
 using System.Text.Json;
-using System.Text.Json.Nodes;
 using Defra.TradeImportsDecisionDeriver.Deriver.Consumers;
 using Defra.TradeImportsDecisionDeriver.Deriver.Interceptors;
 using SlimMessageBus.Host;
